entitiy 이름 제거
