+++ src/main/kotlin/join/famila/persistent/entity/club/Club.kt
@@ -6,8 +6,8 @@ import java.time.LocalDateTime
 import java.time.LocalDateTime.now
 
 @Entity
-@Table(name = "club")
-class ClubEntity(
+@Table
+class Club(
     @Id
     @GeneratedValue(strategy = IDENTITY)
     val id: Long = 0,
@@ -17,23 +17,23 @@ class ClubEntity(
     val introduce: String,
 
     @OneToMany(mappedBy = "club", cascade = [CascadeType.ALL], orphanRemoval = true)
-    val members: List<MemberEntity> = listOf(),
+    val members: List<Member> = listOf(),
 
     @OneToMany(mappedBy = "club", cascade = [CascadeType.ALL], orphanRemoval = true)
-    val posts: List<PostEntity> = listOf(),
+    val posts: List<Post> = listOf(),
 
     @OneToMany(mappedBy = "club", cascade = [CascadeType.ALL], orphanRemoval = true)
-    val pictures: List<PictureEntity> = listOf(),
+    val pictures: List<Picture> = listOf(),
 
     @OneToMany(mappedBy = "club", cascade = [CascadeType.ALL], orphanRemoval = true)
-    val schedules: List<ScheduleEntity> = listOf(),
+    val schedules: List<Schedule> = listOf(),
 
     @CollectionTable(
         name = "category",
         joinColumns = [JoinColumn(name = "club_id", referencedColumnName = "category_id")],
     )
     @ElementCollection
-    val categories: List<CategoryEntity> = listOf(),
+    val categories: List<Category> = listOf(),
 
     val createdAt: LocalDateTime = now(),
