+++ src/main/kotlin/join/famila/persistent/entity/club/Member.kt
@@ -1,0 +1,23 @@
+package join.famila.persistent.entity.club
+
+import jakarta.persistence.*
+import java.time.LocalDateTime
+import java.time.LocalDateTime.now
+
+@Entity
+@Table
+class Member(
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    val id: Long,
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "club_id")
+    val club: Club,
+
+    val userId: Long,
+
+    val createdAt: LocalDateTime = now(),
+
+    val deletedAt: LocalDateTime? = null,
+)
