+++ src/main/kotlin/join/famila/persistent/entity/club/Picture.kt
@@ -5,19 +5,23 @@ import java.time.LocalDateTime
 import java.time.LocalDateTime.now
 
 @Entity
-@Table(name = "member")
-class MemberEntity(
+@Table
+class Picture(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    val id: Long,
+    val id: Long = 0,
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "club_id")
-    val club: ClubEntity,
+    val club: Club,
 
-    val userId: Long,
+    val memberId: Long,
+
+    val link: String,
 
     val createdAt: LocalDateTime = now(),
+
+    val updatedAt: LocalDateTime? = null,
 
     val deletedAt: LocalDateTime? = null,
 )
