+++ src/main/kotlin/join/famila/persistent/entity/club/Schedule.kt
@@ -6,15 +6,15 @@ import java.time.LocalDateTime
 import java.time.LocalDateTime.*
 
 @Entity
-@Table(name = "schedule")
-class ScheduleEntity(
+@Table
+class Schedule(
     @Id
     @GeneratedValue(strategy = IDENTITY)
     val id: Long = 0,
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "club_id")
-    val club: ClubEntity,
+    val club: Club,
 
     val startedAt: LocalDateTime,
